Modified div capture for AdjustmentPanel, set trigger to
Next step: add cookie to detect if user has completed onboarding tutorial
