+++ components/Trade/Advanced/index.tsx
@@ -6,11 +6,6 @@ import styled from 'styled-components';
 import TradingView from './RightPanel';
 import { MARKET } from '@context/OrderContext';
 import { useWeb3 } from '@context/Web3Context/Web3Context';
-import dynamic from 'next/dynamic';
-import { disableBodyScroll, enableBodyScroll } from "body-scroll-lock";
-import { tourConfig } from './TourSteps'
-
-const Tour = dynamic(import('reactour'), { ssr: false });
 
 const TradingPanel = styled.div`
     width: 25%;
@@ -52,7 +47,6 @@ const Advanced: React.FC = styled(({ className }) => {
     const { selectedTracer } = useContext(TracerContext);
     const { order, orderDispatch = () => console.error('Order dispatch not set') } = useContext(OrderContext);
     const [isAdjust] = useState(false);
-    const [isTourOpen, setTourOpen] = useState<boolean>(false);
 
     useEffect(() => {
         if (orderDispatch) {
@@ -72,49 +66,11 @@ const Advanced: React.FC = styled(({ className }) => {
             }
         }
     }, [isAdjust]);
-
-    const closeTour = () => {
-        setTourOpen(false);
-
-        // Reset the elements affected by the tour
-
-        // Show the 'No Position Open' again
-        const positionOverlay = document.querySelector('div[class*="PositionOverlay__StyledOverlay"]') as HTMLElement;
-        if (positionOverlay) {
-            positionOverlay.style.display = 'block';
-        }
-
-        // Disable the 'Close Position' button
-        const closeOrderButton = document.querySelector('button[class*="OrderButtons__CloseOrder"]') as HTMLButtonElement;
-        if (closeOrderButton) {
-            closeOrderButton.disabled = true;
-        }
-    };
-
-    const highlightDots = (e: HTMLDivElement) => {
-        e.addEventListener('click', function () {
-            const navDots: Array<any> = Array.from(document.querySelectorAll('nav[data-tour-elem="navigation"] button'));
-            var currentIndex = 0;
-            // Wait for Reactour to apply styling
-            setTimeout(function(){
-                navDots.map((dot, i) => { 
-                    if(dot.classList.contains('reactour__dot--is-active')){
-                        currentIndex = i;
-                    }
-                });
-                navDots.slice(0, currentIndex).map((dot) => {
-                    dot.classList.add('reactour__dot--is-active');
-                });
-            }, 10);
-        });
-        // Also prevent body scrolling when tour open
-        disableBodyScroll(e);
-    }
     
     return (
         // Remove after testing
         <div className={`container ${className}`}>
-            <TradingPanel onClick={() => setTourOpen(true)}  >
+            <TradingPanel>
                 <MarketSelect account={account ?? ''} />
                 <PlaceOrder selectedTracer={selectedTracer} account={account ?? ''} />
                 <AccountPanel selectedTracer={selectedTracer} account={account ?? ''} order={order} />
@@ -123,19 +79,6 @@ const Advanced: React.FC = styled(({ className }) => {
                 <TradingView selectedTracer={selectedTracer} />
             </RightPanel>
             <Overlay id="trading-overlay" />
-            {/* TODO: detect wallet connect first time, set a cookie */}
-            <Tour
-                onRequestClose={closeTour}
-                steps={tourConfig as Array<any>}
-                maskSpace={0}
-                isOpen={isTourOpen}
-                maskClassName="mask"
-                className="helper"
-                rounded={5}
-                showNumber={false}
-                onAfterOpen={(e) => highlightDots(e)}
-                onBeforeClose={(e) => enableBodyScroll(e)}
-            />
         </div>
     );
 })`
